adafactor

+++ src/train_xla.py
@@ -43,9 +43,9 @@ def _mp_fn(index, args):
         log_print("Syncing model...")
 
         # broadcast with bfloat16 for speed
-        # model = model.to(torch.bfloat16)
+        model = model.to(torch.bfloat16)
         xm.broadcast_master_param(model)
-        # model = model.to(torch.float32)
+        model = model.to(torch.float32)
 
     log_print("Loading data...")
     loader = get_loader(
